Ballistics/Realistic Names - Integrate P90 Compat (#6667)
* Ballistics/Realistic Names - Integrate P90 Compat

* Update CfgWeapons.hpp

* Remove old compat addon

* Update addons/realisticnames/stringtable.xml


* Update addons/realisticnames/CfgWeapons.hpp


* Update addons/ballistics/CfgWeapons.hpp


* Update addons/ballistics/CfgWeapons.hpp
