+++ addons/main/config.cpp
@@ -8,6 +8,7 @@ class CfgPatches {
         requiredVersion = REQUIRED_VERSION;
         requiredAddons[] = {
             "A3_Data_F_Tank_Loadorder",
+            "A3_Data_F_Mod_Loadorder",
             // CBA
             "cba_ui",
             "cba_xeh",
