--- optionals/compat_adr_97/$PBOPREFIX$
@@ -1,1 +1,0 @@
-z\ace\addons\compat_adr_90
