--- optionals/compat_adr_97/CfgMagazines.hpp
@@ -1,7 +1,0 @@
-class CfgMagazines {
-    class CA_Magazine;
-    class 50Rnd_570x28_SMG_03: CA_Magazine {
-        displayName = CSTRING(P90_Mag_Name);
-        descriptionShort = CSTRING(P90_Mag_DescriptionShort);
-    };
-};
