--- optionals/compat_adr_97/meta.cpp
@@ -1,2 +1,0 @@
-protocol = 1;
-publishedid = 773136286;
