--- optionals/compat_adr_97/script_component.hpp
@@ -1,6 +1,0 @@
-#define COMPONENT compat_adr_97
-#define COMPONENT_BEAUTIFIED ADR-97 Compatibility
-
-#include "\z\ace\addons\main\script_mod.hpp"
-
-#include "\z\ace\addons\main\script_macros.hpp"
